refactor(errors): consolidate error hierarchy under lib/errors
Drop unused error and debug modules.
