--- src/lib/errors/configurationerror.js
@@ -1,9 +1,0 @@
-const TronBoxError = require('./tronboxerror');
-
-class ConfigurationError extends TronBoxError {
-  constructor(message) {
-    super(message);
-  }
-}
-
-module.exports = ConfigurationError;
