--- src/lib/errors/taskerror.js
@@ -1,8 +1,0 @@
-const TronBoxError = require('./tronboxerror');
-class TaskError extends TronBoxError {
-  constructor(message) {
-    super(message);
-  }
-}
-
-module.exports = TaskError;
