+++ src/lib/errors/tronBoxError.js
